requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
az-iranian-bank-gateways==1.9.0
blessed==1.20.0
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==4.2.6
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-cleanup==8.1.0
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-picklefield==3.1
django-q==1.3.9
django-smart-selects==1.7.2
django-tinymce==3.7.1
gunicorn==21.2.0
idna==3.6
ippanel==2.0.7
isodate==0.6.1
jinxed==1.2.1
lxml==5.1.0
packaging==23.2
persian-tools==0.0.11
pillow==10.2.0
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pycryptodome==3.20.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2024.1
redis==3.5.3
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.8.19.20240106
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
whitenoise==6.6.0
wrapt==1.16.0
zeep==4.2.1
